add logics

+++ app/Http/Controllers/Api/HouseController.php
@@ -7,11 +7,14 @@ use App\Http\Requests\Api\StoreHouseRequest;
 use App\Http\Requests\Api\UpdateHouseRequest;
 use App\Http\Resources\HouseResource;
 use App\Models\House;
+use Illuminate\Support\Facades\Gate;
 
 class HouseController extends Controller
 {
     public function index()
     {
+        Gate::authorize('viewAny', House::class);
+
         $houses = House::paginate(10);
 
         return HouseResource::collection($houses);
@@ -19,18 +22,32 @@ class HouseController extends Controller
 
     public function store(StoreHouseRequest $request)
     {
-        $house = House::create($request->validated());
+        Gate::authorize('create', House::class);
+
+        $data = $request->validated();
+        
+        $data['user_id'] = auth()->id();
+
+        $house = House::create($data);
 
         return new HouseResource($house);
     }
 
     public function show(House $house)
     {
+        $user = auth()->user();
+
+        Gate::authorize('view', [$house, $user]);
+
         return new HouseResource($house);
     }
 
     public function update(UpdateHouseRequest $request, House $house)
     {
+        $user = auth()->user();
+
+        Gate::authorize('update', [$house, $user]);
+
         $house->update($request->validated());
 
         return new HouseResource($house);
@@ -38,6 +55,10 @@ class HouseController extends Controller
 
     public function destroy(House $house)
     {
+        $user = auth()->user();
+
+        Gate::authorize('delete', [$house, $user]);
+
         $house->delete();
 
         return response()->noContent();
